MFH: Fix compiler warnings
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_3@280806 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/com_dotnet/com_iterator.c
@@ -111,7 +111,7 @@ static int com_iter_move_forwards(zend_object_iterator *iter TSRMLS_DC)
 		}
 	} else {
 		/* safe array */
-		if (I->key >= I->sa_max) {
+		if (I->key >= (ULONG) I->sa_max) {
 			I->key = (ulong)-1;
 			return FAILURE;
 		}
